ML efficiency + unified Nyquist encoder front-end
Training-loop efficiency work (ml-architectures/efficiency-optimization) and a
unified Nyquist temporal front-end shared by all station encoders.

- station_encoders: InputDecimator (model-entry Nyquist decimation, anti-alias FIR);
  PNO SpectralConv1d FFT runs fp32 under bf16 autocast (no half FFT kernel).
- cnn_feature_extractor: SeismicTraceCNN gains a `downsample` param (single same-padded
  strided conv, ceil(L/ds)) so cnn/tcn/pno share one downsample knob; short-input safe.
- seismogram_transformer/train: config-gated perf toggles (ml_perf amp/sdpa/fused_adam/
  compile, ml_encoder input_decimate); fused AdamW + optional torch.compile wiring;
  amplitude_embedding quantile cast to fp32 (bf16-safe under newer torch).
- dataloading/real_noise: in-RAM sim + noise caches (ml_cache).
- axial_transformer/pma_pooling/fused_attention: fused SDPA attention path.
- plotting: drop the dead MTfit import (distributions) + scipy.integrate.simps shim
  (patched_chainconsumer) so the stack imports on modern scipy without MTfit.
- train_NPE: parse ml_encoder.input_decimate / ml_perf; tests for decimator, cnn
  downsample, and per-encoder bf16 amp-safety.

All toggles default-off / byte-identical when unset. Fast gate + ML one-epoch e2e green.

+++ src/seismo_sbi/sbi/compression/ML/amplitude_embedding.py
@@ -187,7 +187,12 @@ class AmplitudeTokenEmbedding(nn.Module):
         B, N, C, T = x.shape
         ax = x.abs().reshape(B, N, C * T)
         peak = ax.amax(dim=-1).clamp_min(self.eps)                                   # (B, N)
-        floor = torch.quantile(ax, self.snr_floor_quantile, dim=-1).clamp_min(self.eps)
+        # torch.quantile requires float/double — under bf16 autocast ax is bf16, so compute
+        # the floor in fp32 (an amplitude statistic; precision here is immaterial) then cast
+        # back. Newer torch rejects bf16 outright; torch 2.0 tolerated it silently.
+        floor = torch.quantile(
+            ax.float(), self.snr_floor_quantile, dim=-1
+        ).to(ax.dtype).clamp_min(self.eps)
         s = peak.log() - floor.log()                                                 # (B, N)
         slope = F.softplus(self.snr_gate_slope)                                      # > 0
         w = torch.sigmoid(slope * (s - self.snr_gate_bias))                          # (B, N)
